problem 15 and 16 solutions

+++ project_euler/problem_4/src/main.rs
@@ -1,12 +1,12 @@
 extern crate prime_utils;
 
 fn main() {
-    let mut palindrome_list: Vec<u64> = Vec::new();
+    let mut palindrome_list: Vec<u128> = Vec::new();
     
     for x in 900..1000 {
         for y in 900..1000 {
-            let mut number_vector: Vec<u64>; 
-            let mut reversed_number_vector: Vec<u64>; 
+            let mut number_vector: Vec<u128>; 
+            let mut reversed_number_vector: Vec<u128>; 
             number_vector = number_to_vector(x*y);
             reversed_number_vector = reverse_vector(&number_vector);
             if palindrome_test(&number_vector, &reversed_number_vector) {
@@ -37,8 +37,8 @@ fn main() {
 
 
 
-fn number_to_vector(mut number_input: u64) -> Vec<u64> {
-   let mut num_vector: Vec<u64> = Vec::new();           // creates new vector
+fn number_to_vector(mut number_input: u128) -> Vec<u128> {
+   let mut num_vector: Vec<u128> = Vec::new();           // creates new vector
    while number_input >= 1 {                            // if number is at least 1, take the last digit and put it into the first spot of the vector
        num_vector.insert(0, number_input % 10 );
        number_input =  number_input / 10;
@@ -47,15 +47,15 @@ fn number_to_vector(mut number_input: u64) -> Vec<u64> {
    num_vector
 }
 
-fn reverse_vector(input_vector: &Vec<u64>) -> Vec<u64> {
-    let mut reversed_vector: Vec<u64> = Vec::new();     // take the element available and put it at the front of the list
+fn reverse_vector(input_vector: &Vec<u128>) -> Vec<u128> {
+    let mut reversed_vector: Vec<u128> = Vec::new();     // take the element available and put it at the front of the list
     for element in input_vector {
         reversed_vector.insert(0, *element);
     }
     reversed_vector
 }
 
-fn palindrome_test(input_vector: &Vec<u64>, reversed_input_vector: &Vec<u64>) -> bool {
+fn palindrome_test(input_vector: &Vec<u128>, reversed_input_vector: &Vec<u128>) -> bool {
     for x in 0..(input_vector.len() / 2 + 1) {
             if input_vector[x] == reversed_input_vector[x] { continue; }
             else {return false; }
